Move site description to figure script
